+++ lib/pages/home_widgets/catalog_list.dart
@@ -1,0 +1,1 @@
+// TODO Implement this library.
