+++ lib/widgets/home_widgets/catalog_header.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../../widgets/themes.dart';
+
+class CatalogHeader extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        "Catalog App".text.xl5.bold.color(MyTheme.darkBluishColor).make(),
+        "Trending products".text.xl2.make(),
+      ],
+    );
+  }
+}
