+++ lib/widgets/home_widgets/catalog_image.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../../widgets/themes.dart';
+
+class CatalogImage extends StatelessWidget {
+  final String image;
+
+  const CatalogImage({Key? key, required this.image}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.network(
+      image,
+    ).box.roundedLg.p8.color(MyTheme.creamColor).make().p16().w40(context);
+  }
+}
